requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
openai
python-dotenv
